More changes to complete r4 of the PCB.
Change the wreath to a new design.  Export it at a higher DPI (600)
before importing it into KiCAD; that was a big problem with the old
rev (which had a 200 DPI export).

Move the "Love Always" around a bit, and a few other tiny tweaks.
No substantial changes since r3.

Update the title blocks, silkscreened copyright notice, and metadata
to r4.

This revision is what's being sent out for the 26Nov18 production run;
in other words, it's what my friends and family will get for
Christmas.  (The software will continue to be changed leading up to
Christmas.)